+++ src/zebra/story.py
@@ -1,0 +1,45 @@
+# -*- coding: utf-8 -*-
+"""Story: The zebra-bra
+"""
+import os
+import sys
+sys.path.append(os.path.join(os.path.dirname(__file__), '../..'))
+sys.path.append('storybuilder')
+
+from storybuilder.builder import world as wd
+from src.zebra import config as cnf
+THM = cnf.THEMES
+
+
+# scenes
+
+# episodes
+# outline
+def story_baseinfo(w: wd.World):
+    return [
+            ]
+
+def story_outline(w: wd.World):
+    return [
+            ]
+
+# main
+def world():
+    w = wd.World("The zebra-bra project")
+    w.set_db(cnf.CHARAS, cnf.STAGES, cnf.DAYS, cnf.ITEMS, cnf.INFOS, cnf.FLAGS,
+            )
+    return w
+
+def story(w: wd.World):
+    return (w.maintitle("ぜぶらぶら"),
+            )
+
+def main(): # pragma: no cover
+    w = world()
+    return w.build(story(w))
+
+
+if __name__ == '__main__':
+    import sys
+    sys.exit(main())
+